% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoscale.R
\name{autoscale}
\alias{autoscale}
\title{Calculate scale for prior slab scales.}
\usage{
autoscale(x, min.x.sd = 1e-04)
}
\arguments{
\item{x}{A design matrix.}

\item{min.x.sd}{The minimum acceptable standard deviation of each column.}
}
\value{
A vector of scale values.

A vector of scaled values.
}
\description{
The standard deviations for each column in the design matrix are to create
scaling values for the slab prior scale.
}
\note{
This function is taken from the R package \code{BhGLM}.
}
