% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_bh3.R
\name{cv.bh3}
\alias{cv.bh3}
\title{Cross Validation for Bayesian Models or Elastic Net}
\usage{
cv.bh3(object, nfolds = 10, foldid = NULL, ncv = 1, verbose = TRUE)
}
\arguments{
\item{object}{ 
  a fitted object. 
  }

\item{nfolds}{
  number of folds(groups) into which the data should be split to estimate the cross-validation prediction error. 
  default is 10. Although nfolds can be as large as the sample size (leave-one-out CV), it is not recommended for large datasets. 
  }

\item{foldid}{
  an optional vector (if ncv = 1) or matrix (if ncv > 1) of values between 1 and nfolds identifying what fold each observation is in. 
  If supplied, nfolds can be missing.If \code{foldid = NULL}, \code{nfolds} subsets will be generated randomly.  
  }

\item{ncv}{
  repeated number of cross-validation.  
  }

\item{verbose}{
  logical. If \code{TRUE}, print out computational time and progress.
  }
}
\description{
Several alterations to \code{cv.bh()} were necessary to ensure that \code{update()} works in
the functions \code{compare_ssnet()}. The arguments and functionality
are the same as \code{cv.bh()}. See \code{\link[BhGLM]{cv.bh}} for details.
}
\note{
The package \code{pROC} will not calculate the AUC when a fold does does not have
at least one observation of each level. This can largely be avoided by selecting the number
of folds so that such circumstances are rare. When such does occur, the current result is
to assign AUC <- NA.
}
