% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_classify.R
\name{eval_classify}
\alias{eval_classify}
\title{Extract True/False Positive/Negative Measures After Classification}
\usage{
eval_classify(num.class = 2, classify.rule, beta.hat, x, y, family)
}
\arguments{
\item{num.class}{A scalar value defining the number of classes. Default value is two.}

\item{classify.rule}{A vector of thresholds that determine assigned classes.}

\item{beta.hat}{A vector of parameter estimates from some model fit.}

\item{x}{A (numeric) design matrix whose number of columns is equal to \code{length(beta.hat) - 1} and whose number
of rows is equal to \code{length(y)}. Strictly speaking, \code{x} may be a data frame, but categorical variables
must have already been transformed into alternative coding schemes; e.g., dummy coding.}

\item{y}{A numeric vector of outcome measurements. The length of this vector must equal \code{nrow(x)}.}

\item{family}{One of \code{"gaussian", "gamma", "poisson", "binomial"}, which determines the distribution
from which \code{y} is assumed to be drawn.}

\item{make.x0}{Logical. When \code{TRUE}, creates a column of ones for the intercept. Default is \code{TRUE}.}

\item{gamma.link}{Defines the link function for the mean in gamma GLM'. One of \code{c("inverse", "log", "identity")}.}
}
\value{
A data frame with a single row containing columns for sensitivity, specificity, ppv, npv, and accuracy,
where ppv is positive predictive value and npv is negative predictive value. When there are 3 or more classes,
only accuracy is returned.
}
\description{
After fitting GLM's with, set a classification  rule, and obtain measures for classification effectiveness.
}
