% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_iar.R
\name{format_iar}
\alias{format_iar}
\title{Format IAR data for ssnet}
\usage{
format_iar(adjmat = NULL, im.res = NULL, x = NULL)
}
\arguments{
\item{adjmat}{A data.frame or matrix containing a "sparse" representation of the neighbor relationships. The first
column should contain a numerical index for a given location. Each index will be repeated in this column for
every neighbor it has. The indices for the location's neighbors are then specified in the second column. An optional
third column specifies weights. If no third column is specified, then equal weights are assumed.}

\item{im.res}{A 2-element vector where the first argument is the number of "rows" and the second argument
is the number of "columns" in each subject's "image". Default is \code{NULL}.}

\item{x}{Design, or input, matrix, of dimension nobs x nvars; each row is an observation vector. It is recommended that
\code{x} have user-defined column names for ease of identifying variables.}
}
\description{
Given an adjacency matrix, formats data for using IAR priors to model spatial structure. Otherwise, assumes
default structure where neighbors are defined as variables directly above or to the side of a specified
variable location.
}
