% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_foldid.R
\name{generate.foldid}
\alias{generate.foldid}
\title{Generate Folds for K-Fold Cross Validation}
\usage{
generate.foldid(nobs, nfolds = 10, foldid = NULL, ncv = 1)
}
\description{
Unchanged from \code{BhGLM}, but was not in NAMESPACE, so reproduced here.
}
