% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet_measure.R
\name{glmnet_measure}
\alias{glmnet_measure}
\title{Extract Measures of Model Fit from GLMNET Fit}
\usage{
glmnet_measure(
  cv.model,
  x = NULL,
  y = NULL,
  lambda.criteria = "lambda.min",
  type.measure = c("default", "mse", "deviance", "class", "auc", "mae", "C"),
  family = "gaussian",
  alpha = 1
)
}
\arguments{
\item{cv.model}{A model fit by \code{glmnet} function \code{cv.glmnet()}.}

\item{x}{input matrix, of dimension nobs x nvars; each row is an observation
vector. Can be in sparse matrix format (inherit from class
\code{"sparseMatrix"} as in package \code{Matrix}).
Requirement: \code{nvars >1}; in other words, \code{x} should have 2 or more columns.}

\item{y}{response variable. Quantitative for \code{family="gaussian"}, or
\code{family="poisson"} (non-negative counts). For \code{family="binomial"}
should be either a factor with two levels, or a two-column matrix of counts
or proportions (the second column is treated as the target class; for a
factor, the last level in alphabetical order is the target class). For
\code{family="multinomial"}, can be a \code{nc>=2} level factor, or a matrix
with \code{nc} columns of counts or proportions. For either
\code{"binomial"} or \code{"multinomial"}, if \code{y} is presented as a
vector, it will be coerced into a factor. For \code{family="cox"}, preferably
a \code{Surv} object from the survival package: see Details section for
more information. For \code{family="mgaussian"}, \code{y} is a matrix
of quantitative responses.}

\item{lambda.criteria}{Determines the model selection criteria. When \code{"lambda.min"} the
final model is selected based on the penalty that minimizes the measure given in \code{type.measure}.
When \code{"lambda.1se"} the final model is selected based on the smallest value of lambda that
is within one standard error of the minimal measure given in \code{type.measure}.}

\item{type.measure}{loss to use for cross-validation. Currently five
options, not all available for all models. The default is
\code{type.measure="deviance"}, which uses squared-error for gaussian models
(a.k.a \code{type.measure="mse"} there), deviance for logistic and poisson
regression, and partial-likelihood for the Cox model.
\code{type.measure="class"} applies to binomial and multinomial logistic
regression only, and gives misclassification error.
\code{type.measure="auc"} is for two-class logistic regression only, and
gives area under the ROC curve. \code{type.measure="mse"} or
\code{type.measure="mae"} (mean absolute error) can be used by all models
except the \code{"cox"}; they measure the deviation from the fitted mean to
the response.
\code{type.measure="C"} is Harrel's concordance measure, only available for \code{cox} models.}

\item{family}{Either a character string representing
one of the built-in families, or else a \code{glm()} family object. For more
information, see Details section below or the documentation for response
type (above).}

\item{alpha}{The elasticnet mixing parameter, with \eqn{0\le\alpha\le 1}.
The penalty is defined as
\deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} is the
lasso penalty, and \code{alpha=0} the ridge penalty.}
}
\value{
A data frame containing measures of model fitness.
}
\description{
Extracts measures of model fitness from a \code{glmnet} object.
}
\note{
This function is primarily used within \code{\link[ssnet]{compare_ssnet}}, but perhaps could be useful elsewhere.
}
\examples{
## generate data (no intercept)
set.seed(4799623)
cn <- c()
for (i in 1:100) cn[i] <- paste0("x", i)
tb <- rbinom(100, 1, 0.05)
tx <- matrix(rnorm(10000), nrow = 100, ncol = 100,
             dimnames = list(1:100, cn))
ty <- tx \%*\% tb + rnorm(100)

## fit model and export model fit stats
cv1 <- cv.glmnet(x = tx, y = ty, family = "gaussian", alpha = 1)
glmnet_measure(cv1, x = tx, y = ty)
}
