% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grouping.R
\name{Grouping}
\alias{Grouping}
\title{Prepare variables for group-based inference.}
\usage{
Grouping(all.var, group)
}
\arguments{
\item{all.var}{A vector containing the names of the variables in the design
matrix.}

\item{group}{A numeric vector, or an integer, or a list indicating the
groups of predictors. If \code{group = NULL}, all the predictors form a
single group. If \code{group = K}, the predictors are evenly divided into
groups each with K predictors. If group is a numeric vector, it defines
groups as follows: Group 1: \code{(group[1]+1):group[2]}, Group 2:
\code{(group[2]+1):group[3]}, Group 3: \code{(group[3]+1):group[4]}, ...
If group is a list of variable names, \code{group[[k]]} includes variables
in the k-th group. The mixture double-exponential prior is only used for
grouped predictors. For ungrouped predictors, the prior is
double-exponential with scale \code{ss[2]} and mean 0.}
}
\value{
A list.
}
\description{
This function uses the user-defined groups to properly prepare the variables
or group-specific estimation/inference.
}
\note{
This function is a modified version of the function \code{Grouping}
from the R package \code{BhGLM}.
}
