% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_raw.R
\name{measure_bh_raw}
\alias{measure_bh_raw}
\title{Obtain Measures of Model Fitness}
\usage{
measure_bh_raw(object, new.x, new.y, new.offset)
}
\arguments{
\item{object}{A fitted object..}

\item{new.x}{A data frame or matrix of new values for variables used in object.
For an object from \code{bglm}, \code{bpolr} or \code{bcoxph}, it is data frame;
for object from \code{glmNet} or \code{bmlasso}, it is matrix.}

\item{new.y}{A vector of new response values corresponding to new.x. If \code{new.x}
or \code{new.y} are omitted, the fitted linear predictors are used for prediction.}

\item{new.offset}{A data frame or vector of offset values for new data points.
If \code{new.x} includes offset, do not need to set \code{new.offset}.}
}
\value{
A vector.
}
\description{
Obtain Measures of Model Fitness
}
\note{
This function is taken directly from \code{measure.bh} in \code{BhGLM}, with the modification that
measures are no longer rounded.
}
