% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungeCARdata4stan.R
\name{mungeCARdata4stan}
\alias{mungeCARdata4stan}
\title{Prepare necessary components of the IAR model for use in \code{stan}.}
\usage{
mungeCARdata4stan(adjBUGS, numBUGS)
}
\arguments{
\item{adjBUGS}{A vector of neighbor indices.}

\item{numBUGS}{A vector containing the number of neighbors for each point.}
}
\value{
Returns a list where \code{J} is the number predictors, \code{J_edges} is the number
of edges in connected graph representing the spatial data, and \code{node1, node2} are required
to fit the model without explicitly specifying an adjacency matrix. This function is essentially
unchanged from Mitzi Morris's code \insertCite{Morris:2017,Morris:2019}{ssnet}.
}
\description{
The underlying `stan` code can work without explicitly calling or building the adjacency
matrix, but to do so requires some processing, which is done here.
}
\references{
\insertRef{Morris:2017}{ssnet}

\insertRef{Morris:2019}{ssnet}
}
