% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssnet-package.R
\docType{package}
\name{ssnet-package}
\alias{ssnet-package}
\title{The 'ssnet' package.}
\description{
Fit Spike-and-Slab Elastic Net Models
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package
version 2.19.3. https://mc-stan.org
}
