% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scale_p.R
\name{update_scale_p}
\alias{update_scale_p}
\title{Update conditional expectation of prior probabilities of inclusion at each iteration.}
\usage{
update_scale_p(b0, ss, theta)
}
\arguments{
\item{b0}{A vector of current parameter estimates.}

\item{ss}{A vector of spike and slab prior scales, respectively.}

\item{theta}{The current estimate of prior probabilities of inclusion.}
}
\value{
A list whose first element is a vector of updated scale parameters for each
parameter and whose second element is a vector of updated conditional expectations
of prior probabilities of model inclusion.
}
\description{
Update conditional expectation of prior probabilities of inclusion at each iteration.
}
\note{
This function is taken unchanged (except for the name) from the R package \code{BhGLM}.
}
