% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_foldid.R
\name{generate.foldid}
\alias{generate.foldid}
\title{Generate Folds for K-Fold Cross Validation}
\usage{
generate.foldid(nobs, nfolds = 10, foldid = NULL, ncv = 1, fold.seed = NULL)
}
\arguments{
\item{nobs}{The number of observations.}

\item{nfolds}{The number of folds to generate.}

\item{foldid}{An array of fold identifiers. If not supplied by the user,
generated within function.}

\item{ncv}{
  repeated number of cross-validation.  
  }

\item{fold.seed}{An integer that sets the seed for generating folds.}
}
\description{
Unchanged from \code{BhGLM}, but was not in NAMESPACE, so reproduced here.
}
