% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_foldid.R
\name{generate.foldid}
\alias{generate.foldid}
\title{Generate Folds for K-Fold Cross Validation}
\usage{
generate.foldid(nobs, nfolds = 10, foldid = NULL, ncv = 1)
}
\arguments{
\item{nobs}{The number of observations.}

\item{nfolds}{ 
  These arguments are the same as in the functions \code{\link{glmnet}} and \code{\link{cv.glmnet}} in the package \bold{glmnet}.
  }

\item{foldid}{An array of fold identifiers. If not supplied by the user, generated within function.}

\item{ncv}{
  repeated number of cross-validation.  
  }
}
\description{
Unchanged from \code{BhGLM}, but was not in NAMESPACE, so reproduced here.
}
