% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_q2_iar.R
\name{max_q2_iar}
\alias{max_q2_iar}
\title{Update prior probabilities of model inclusion at each iteration.}
\usage{
max_q2_iar(
  iar.data,
  p,
  opt.algorithm = "LBFGS",
  tau.prior = "none",
  tau.manual = NULL,
  p.bound = c(0.01, 0.99),
  stan_manual = NULL
)
}
\arguments{
\item{iar.data}{A list of output from \code{\link{mungeCARdata4stan}} that contains the necessary
inputs for the IAR prior.}

\item{p}{A vector containing the prior probabilities of inclusion for each parameter at the current iteration.}

\item{opt.algorithm}{One of the optimization algorithms available from \code{optimizing}.}

\item{tau.prior}{One of \code{c("none", "manual", "cauchy")}. This argument determines the precision
parameter in the Conditional Autoregressive model for the (logit of) prior inclusion probabilities.
When \code{"none"}, the precision is set to 1; when "manual", the precision is manually entered by the
user (and still not random); when \code{"cauchy"}, the inverse precision is assumed to follow a Cauchy
distribution with mean 0 and scale 2.5. Note that at this stage of development, only the \code{"none"} option
has been extensively tested, so the other options should be used with caution.}

\item{tau.manual}{When \code{tau.prior = "manual"}, use this argument to specify a common precision
parameter.}

\item{p.bound}{A vector defining the lower and upper boundaries for the probabilities of inclusion
in the model, respectively. Defaults to \code{c(0.01, 0.99)}.}

\item{stan_manual}{A \code{stan_model} that is manually specified.}
}
\value{
A vector containing updated probabilities of inclusion.
}
\description{
Uses \code{stan} to optimize the term in the EM algorithm that depends on the prior probabilities
of inclusion in the model, \eqn{p_j}.
}
\note{
This function borrows from the work of Mitzi Morris, who describes how to fit an intrinsic autoregression
in \code{stan} \insertCite{Morris:2017,Morris:2019}{ssnet}.
}
\references{
\insertRef{Morris:2017}{ssnet}

\insertRef{Morris:2019}{ssnet}

\insertRef{Tang:2017}{ssnet}
}
