% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_q2_iar.R
\name{max_q2_iar}
\alias{max_q2_iar}
\title{Update prior probabilities of model inclusion at each iteration.}
\usage{
max_q2_iar(
  iar.data,
  p,
  opt.algorithm = "LBFGS",
  tau.prior = "none",
  p.bound = c(0.01, 0.99),
  stan_manual = NULL,
  stan_local = FALSE
)
}
\arguments{
\item{iar.data}{A list of output from \code{\link{mungeCARdata4stan}} that contains the necessary
inputs for the IAR prior.}

\item{p}{A vector containing the prior probabilities of inclusion for each parameter at the current iteration.}

\item{opt.algorithm}{One of the optimization algorithms available from \code{optimizing}.}

\item{tau.prior}{One of \code{c("none", "manual", "cauchy")}. This argument determines the precision
parameter in the Conditional Autoregressive model for the (logit of) prior inclusion probabilities.
When \code{"none"}, the precision is set to 1; when "manual", the precision is manually entered by the
use; when \code{"cauchy"}, the inverse precision is assumed to follow a Cauchy distribution with mean 0 and
scale 2.5.}

\item{p.bound}{A vector defining the lower and upper boundaries for the probabilities of inclusion
in the model, respectively. Defaults to \code{c(0.01, 0.99)}.}

\item{stan_manual}{A \code{stan_model} that is manually specified.}

\item{stan_local}{Logical. Defaults to \code{FALSE}, but when \code{TRUE}, uses locally stored \code{stan} files.
This option will eventually be removed once the package is more stable.}
}
\value{
A vector containing updated probabilities of inclusion.
}
\description{
Uses \code{stan} to optimize the term in the EM algorithm that depends on the prior probabilities
of inclusion in the model, \eqn{p_j}.
}
\note{
This function borrows from the work of Mitzi Morris, who describes how to fit an intrisic autoregression
in \code{stan} \insertCite{Morris:2017,Morris:2019}{ssnet}.
}
\examples{
## build adjacency matrix
adjmat10x10 <- proximity_builder(im.res = c(10, 10), type = "sparse")
model_info10x10 <- mungeCARdata4stan(adjmat10x10$nb.index,
                                     table(adjmat10x10$location.index))

## let function call stan model
mq2 <- max_q2_iar(iar.data = model_info10x10, p = rbeta(100, 1, 1))

## specify stan model ahead of time
sm <- stan_model(file = "C:/Users/Justin/Documents/BST/Dissertation_in_Latex/stan models/iar_incl_prob_notau.stan")
mq2 <- max_q2_iar(iar.data = model_info10x10, p = rbeta(100, 1, 1), stan_manual = sm)
}
\references{
\insertRef{Morris:2017}{ssnet}

\insertRef{Morris:2019}{ssnet}

\insertRef{Tang:2017}{ssnet}
}
