% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_scale_p.R
\name{update_scale_p}
\alias{update_scale_p}
\title{Update conditional expectation of prior probabilities of inclusion at each
iteration.}
\usage{
update_scale_p(b0, ss, theta, alpha)
}
\arguments{
\item{b0}{A vector of current parameter estimates.}

\item{ss}{A vector of spike and slab prior scales, respectively.}

\item{theta}{The current estimate of prior probabilities of inclusion.}

\item{alpha}{A scalar value between 0 and 1 determining the compromise
between the Ridge and Lasso models. Whenv\code{alpha = 1} reduces to the
Lasso, and when \code{alpha = 0} reduces to Ridge.}
}
\value{
A list whose first element is a vector of updated scale parameters
for eachvparameter and whose second element is a vector of updated
conditional expectationsvof prior probabilities of model inclusion.
}
\description{
Update conditional expectation of prior probabilities of inclusion at each
iteration.
}
\note{
This function is a modified version of \code{update_scale_p()} from
the R package \code{BhGLM}.
}
